Moved mixpanel connection to frontend
